closes #1 add auth, update screenshots

+++ app/Http/Controllers/AdvertisementController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Advertisement;
 
@@ -12,6 +13,11 @@ class AdvertisementController extends Controller
         'price' => 'required|numeric|between:0,99999.99',
         'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
     ];
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
 
     public function index()
     {
@@ -55,14 +61,23 @@ class AdvertisementController extends Controller
             abort(404);
         }
 
+        if (Auth::id() !== $ad->user_id) {
+            abort(403);
+        }
+
         return view('advertisement.edit', ['ad' => $ad]);
     }
 
     public function update(Request $request, $id)
     {
+        $ad = Advertisement::find($id);
+
+        if (Auth::id() !== $ad->user_id) {
+            abort(403);
+        }
+
         $request->validate(self::FORM_VALIDATION_RULES);
 
-        $ad = Advertisement::find($id);
         $ad->update($request->all());
         $ad->updateImage($request->image);
 
@@ -74,14 +89,26 @@ class AdvertisementController extends Controller
     public function destroy($id)
     {
         $ad = Advertisement::find($id);
+
         if (!$ad) {
-            return redirect(route('advertisement.index'));
+            return redirect(route('advertisement.admin'));
+        }
+
+        if (Auth::id() !== $ad->user_id) {
+            abort(403);
         }
 
         $ad->delete();
 
-        return redirect(route('advertisement.index'))
+        return redirect(route('advertisement.admin'))
             ->with('message_type', 'success')
             ->with('message', 'You have successfully deleted advertisement.');
+    }
+
+    public function admin()
+    {
+        $ads = Advertisement::where('user_id', Auth::id())->get();
+
+        return view('advertisement.admin', ['ads' => $ads]);
     }
 }
